+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.7.1 - soon
 ### Fixed
 - `-f diff` no longer claims that it found more issues when it didn't
+- SC2154 triggers for builtins called with `builtin`
 
 ### Added
 - SC2254: Suggest quoting expansions in case statements